backend/test_etl_mock.py
import pytest
from unittest.mock import MagicMock, patch
import pandas as pd

# We want to test the transformation logic without hitting a real DB or File System

def test_clean_currency_logic():
    """
    Test that our SQL logic for cleaning currency works as expected.
    In a real London School TDD approach, we'd test the 'Behavior' 
    of the transformation function.
    """
    # Simulation of what DuckDB does
    raw_data = ["$100,000", "$50.00", "0"]
    cleaned = [float(x.replace('$', '').replace(',', '')) for x in raw_data]
    
    assert cleaned[0] == 100000.0
    assert cleaned[1] == 50.0
    assert cleaned[2] == 0.0

def test_etl_structure():
    """
    Verify that the ETL script calls the uploader with the correct schema.
    """
    # This is a 'Collaboration Test' (London School)
    # We mock the 'supabase' client and verify 'upsert' is called.
    
    mock_supabase = MagicMock()
    mock_table = MagicMock()
    mock_supabase.table.return_value = mock_table
    
    data = [{"id": 1, "val": 10}]
    
    # Simulated function call
    mock_table.upsert(data).execute()
    
    # Assert interaction
    mock_supabase.table.assert_called_with("fact_layoffs") # Hypothetical
    mock_table.upsert.assert_called_with(data)
