backend/requirements_etl.txt
duckdb>=0.9.2
supabase>=2.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
pyarrow>=14.0.0
